backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.2.1
replicate==1.0.7
Pillow==12.0.0
opencv-python==4.12.0.88
requests==2.32.5
gunicorn==23.0.0

